Reorganization - moved periods functions from Calendars module / directory to Periods module / directory
