+++ lib/mehr_schulferien/periods.ex
@@ -5,8 +5,188 @@ defmodule MehrSchulferien.Periods do
 
   import Ecto.Query, warn: false
 
-  alias MehrSchulferien.Calendars.{BridgeDayPeriod, DateHelpers, Period}
+  alias MehrSchulferien.Calendars
+  alias MehrSchulferien.Calendars.{DateHelpers, HolidayOrVacationType}
+  alias MehrSchulferien.Periods.{BridgeDayPeriod, Period}
   alias MehrSchulferien.Repo
+
+  @doc """
+  Returns the list of periods.
+  """
+  def list_periods do
+    Repo.all(Period)
+  end
+
+  @doc """
+  Returns a list of previous periods for a federal_state.
+  """
+  def list_previous_periods(federal_state, holiday_or_vacation_type) do
+    today = DateHelpers.today_berlin()
+
+    from(p in Period,
+      where:
+        p.location_id == ^federal_state.id and
+          p.holiday_or_vacation_type_id == ^holiday_or_vacation_type.id and
+          p.ends_on < ^today,
+      order_by: [desc: p.starts_on]
+    )
+    |> Repo.all()
+    |> Repo.preload([:holiday_or_vacation_type])
+  end
+
+  @doc """
+  Returns a list of current and future periods for a federal_state.
+  """
+  def list_current_and_future_periods(federal_state, holiday_or_vacation_type) do
+    today = DateHelpers.today_berlin()
+
+    from(p in Period,
+      where:
+        p.location_id == ^federal_state.id and
+          p.holiday_or_vacation_type_id == ^holiday_or_vacation_type.id and
+          p.ends_on >= ^today,
+      order_by: p.starts_on
+    )
+    |> Repo.all()
+    |> Repo.preload([:holiday_or_vacation_type])
+  end
+
+  @doc """
+  Gets a single period.
+
+  Raises `Ecto.NoResultsError` if the Period does not exist.
+  """
+  def get_period!(id), do: Repo.get!(Period, id)
+
+  @doc """
+  Creates a period.
+  """
+  def create_period(%{"holiday_or_vacation_type_id" => holiday_or_vacation_type_id} = attrs)
+      when not is_nil(holiday_or_vacation_type_id) do
+    %HolidayOrVacationType{
+      id: id,
+      default_html_class: html_class,
+      default_is_listed_below_month: is_listed_below_month,
+      default_is_public_holiday: is_public_holiday,
+      default_is_school_vacation: is_school_vacation,
+      default_is_valid_for_everybody: is_valid_for_everybody,
+      default_is_valid_for_students: is_valid_for_students,
+      default_religion_id: religion_id,
+      default_display_priority: display_priority
+    } = Calendars.get_holiday_or_vacation_type!(holiday_or_vacation_type_id)
+
+    attrs =
+      Map.merge(
+        %{
+          "holiday_or_vacation_type_id" => id,
+          "html_class" => html_class,
+          "is_listed_below_month" => is_listed_below_month,
+          "is_public_holiday" => is_public_holiday,
+          "is_school_vacation" => is_school_vacation,
+          "is_valid_for_everybody" => is_valid_for_everybody,
+          "is_valid_for_students" => is_valid_for_students,
+          "religion_id" => religion_id,
+          "display_priority" => display_priority
+        },
+        attrs
+      )
+
+    %Period{}
+    |> Period.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  def create_period(%{holiday_or_vacation_type_id: holiday_or_vacation_type_id} = attrs)
+      when not is_nil(holiday_or_vacation_type_id) do
+    attrs = for {key, val} <- attrs, into: %{}, do: {to_string(key), val}
+    create_period(attrs)
+  end
+
+  def create_period(attrs) do
+    %Period{}
+    |> Period.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a period.
+  """
+  def update_period(%Period{} = period, attrs) do
+    period
+    |> Period.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a period.
+  """
+  def delete_period(%Period{} = period) do
+    Repo.delete(period)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking period changes.
+  """
+  def change_period(%Period{} = period) do
+    Period.changeset(period, %{})
+  end
+
+  @doc """
+  Finds all the holiday periods for a certain date.
+  """
+  def find_all_periods(periods, date) do
+    Enum.filter(periods, &is_holiday?(&1, date))
+  end
+
+  @doc """
+  Returns the next schoolday (the next day that is not a holiday).
+  """
+  def find_next_schoolday([], _), do: nil
+
+  def find_next_schoolday([first | rest], date) do
+    if is_holiday?(first, date) do
+      new_date = Date.add(first.ends_on, 1)
+      find_next_schoolday(rest, new_date)
+    else
+      if check_ends_on(date, first) do
+        date
+      else
+        find_next_schoolday(rest, date)
+      end
+    end
+  end
+
+  defp is_holiday?(period, date) do
+    case Date.compare(date, period.starts_on) do
+      :lt -> false
+      :eq -> true
+      :gt -> check_ends_on(date, period)
+    end
+  end
+
+  defp check_ends_on(date, first) do
+    case Date.compare(date, first.ends_on) do
+      :gt -> nil
+      _ -> first
+    end
+  end
+
+  @doc """
+  Returns the holiday periods for a certain date's month.
+  """
+  def find_periods_by_month(_date, []), do: []
+
+  def find_periods_by_month(date, [first | rest]) do
+    if DateHelpers.compare_by_month(date, first.starts_on) == :lt do
+      []
+    else
+      if DateHelpers.compare_by_month(date, first.ends_on) == :gt do
+        find_periods_by_month(date, rest)
+      else
+        [first | find_periods_by_month(date, rest)]
+      end
+    end
+  end
 
   @doc """
   Returns the periods for a certain date range.
