+++ lib/mehr_schulferien_web/controllers/api/v2/period_controller.ex
@@ -1,15 +1,15 @@
 defmodule MehrSchulferienWeb.Api.V2.PeriodController do
   use MehrSchulferienWeb, :controller
 
-  alias MehrSchulferien.Calendars
+  alias MehrSchulferien.Periods
 
   def index(conn, _params) do
-    periods = Calendars.list_periods()
+    periods = Periods.list_periods()
     render(conn, "index.json", periods: periods)
   end
 
   def show(conn, %{"id" => id}) do
-    period = Calendars.get_period!(id)
+    period = Periods.get_period!(id)
     render(conn, "show.json", period: period)
   end
 end
