+++ lib/mehr_schulferien_web/controllers/period_controller.ex
@@ -1,22 +1,21 @@
 defmodule MehrSchulferienWeb.PeriodController do
   use MehrSchulferienWeb, :controller
 
-  alias MehrSchulferien.Calendars
-  alias MehrSchulferien.Calendars.Period
+  alias MehrSchulferien.{Periods, Periods.Period}
   alias MehrSchulferien.Repo
 
   def index(conn, _params) do
-    periods = Calendars.list_periods() |> Repo.preload([:location, :holiday_or_vacation_type])
+    periods = Periods.list_periods() |> Repo.preload([:location, :holiday_or_vacation_type])
     render(conn, "index.html", periods: periods)
   end
 
   def new(conn, _params) do
-    changeset = Calendars.change_period(%Period{})
+    changeset = Periods.change_period(%Period{})
     render(conn, "new.html", changeset: changeset)
   end
 
   def create(conn, %{"period" => period_params}) do
-    case Calendars.create_period(period_params) do
+    case Periods.create_period(period_params) do
       {:ok, period} ->
         conn
         |> put_flash(:info, "Period created successfully.")
@@ -28,20 +27,20 @@ defmodule MehrSchulferienWeb.PeriodController do
   end
 
   def show(conn, %{"id" => id}) do
-    period = Calendars.get_period!(id)
+    period = Periods.get_period!(id)
     render(conn, "show.html", period: period)
   end
 
   def edit(conn, %{"id" => id}) do
-    period = Calendars.get_period!(id)
-    changeset = Calendars.change_period(period)
+    period = Periods.get_period!(id)
+    changeset = Periods.change_period(period)
     render(conn, "edit.html", period: period, changeset: changeset)
   end
 
   def update(conn, %{"id" => id, "period" => period_params}) do
-    period = Calendars.get_period!(id)
+    period = Periods.get_period!(id)
 
-    case Calendars.update_period(period, period_params) do
+    case Periods.update_period(period, period_params) do
       {:ok, period} ->
         conn
         |> put_flash(:info, "Period updated successfully.")
@@ -53,8 +52,8 @@ defmodule MehrSchulferienWeb.PeriodController do
   end
 
   def delete(conn, %{"id" => id}) do
-    period = Calendars.get_period!(id)
-    {:ok, _period} = Calendars.delete_period(period)
+    period = Periods.get_period!(id)
+    {:ok, _period} = Periods.delete_period(period)
 
     conn
     |> put_flash(:info, "Period deleted successfully.")
