+++ lib/mehr_schulferien_web/controllers/school_controller.ex
@@ -3,7 +3,8 @@ defmodule MehrSchulferienWeb.SchoolController do
 
   import MehrSchulferienWeb.Authorize
 
-  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Calendars.Period, Locations}
+  alias MehrSchulferien.{Calendars, Calendars.DateHelpers, Locations}
+  alias MehrSchulferien.{Periods, Periods.Period}
   alias MehrSchulferienWeb.ControllerHelpers, as: CH
   alias MehrSchulferienWeb.Email
 
@@ -20,7 +21,7 @@ defmodule MehrSchulferienWeb.SchoolController do
         "Schulschließung wegen der COVID-19-Pandemie (Corona)"
       )
 
-    changeset = Calendars.change_period(%Period{})
+    changeset = Periods.change_period(%Period{})
 
     render(conn, "new.html",
       changeset: changeset,
@@ -38,7 +39,7 @@ defmodule MehrSchulferienWeb.SchoolController do
         "school_slug" => school_slug,
         "period" => period_params
       }) do
-    case Calendars.create_period(period_params) do
+    case Periods.create_period(period_params) do
       {:ok, period} ->
         Email.period_added_notification(period)
