+++ test/mehr_schulferien/calendars_test.exs
@@ -4,7 +4,7 @@ defmodule MehrSchulferien.CalendarsTest do
   import MehrSchulferien.Factory
 
   alias MehrSchulferien.Calendars
-  alias MehrSchulferien.Calendars.{HolidayOrVacationType, Period, Religion}
+  alias MehrSchulferien.Calendars.{HolidayOrVacationType, Religion}
 
   describe "religions" do
     @valid_attrs %{
@@ -152,128 +152,6 @@ defmodule MehrSchulferien.CalendarsTest do
 
       assert %Ecto.Changeset{} =
                Calendars.change_holiday_or_vacation_type(holiday_or_vacation_type)
-    end
-  end
-
-  describe "periods" do
-    @valid_attrs %{
-      "created_by_email_address" => "george@example.com",
-      "ends_on" => ~D[2010-04-20],
-      "starts_on" => ~D[2010-04-17]
-    }
-    @update_attrs %{
-      "html_class" => "white",
-      "is_public_holiday" => true
-    }
-    @invalid_attrs %{
-      "created_by_email_address" => "george@example.com",
-      "ends_on" => ~D[2010-04-20],
-      "starts_on" => ~D[2010-04-27]
-    }
-
-    test "list_periods/0 returns all periods" do
-      period = insert(:period)
-      assert [period_1] = Calendars.list_periods()
-      assert period.id == period_1.id
-    end
-
-    test "get_period!/1 returns the period with given id" do
-      period = insert(:period)
-      assert period_1 = Calendars.get_period!(period.id)
-      assert period.id == period_1.id
-    end
-
-    test "create_period/1 with valid data creates a period" do
-      location = insert(:federal_state)
-      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
-
-      valid_attrs =
-        Map.merge(@valid_attrs, %{
-          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
-          "location_id" => location.id
-        })
-
-      assert {:ok, %Period{} = period} = Calendars.create_period(valid_attrs)
-
-      assert period.created_by_email_address == "george@example.com"
-      assert period.ends_on == ~D[2010-04-20]
-      assert period.starts_on == ~D[2010-04-17]
-      assert period.html_class == holiday_or_vacation_type.default_html_class
-      assert period.is_public_holiday == holiday_or_vacation_type.default_is_public_holiday
-      assert period.religion_id == holiday_or_vacation_type.default_religion_id
-    end
-
-    test "create_period/1 with attrs overriding holiday_or_vacation_type defaults" do
-      location = insert(:federal_state)
-      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
-
-      valid_attrs =
-        Map.merge(@valid_attrs, %{
-          "location_id" => location.id,
-          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
-          "is_public_holiday" => !holiday_or_vacation_type.default_is_public_holiday
-        })
-
-      assert {:ok, %Period{} = period} = Calendars.create_period(valid_attrs)
-      assert period.created_by_email_address == "george@example.com"
-      assert period.is_public_holiday != holiday_or_vacation_type.default_is_public_holiday
-    end
-
-    test "create_period/1 with invalid data returns error changeset" do
-      location = insert(:federal_state)
-      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
-
-      invalid_attrs =
-        Map.merge(@invalid_attrs, %{
-          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
-          "location_id" => location.id
-        })
-
-      assert {:error, %Ecto.Changeset{} = changeset} = Calendars.create_period(invalid_attrs)
-
-      assert %{starts_on: ["starts_on should be less than or equal to ends_on"]} =
-               errors_on(changeset)
-
-      invalid_attrs = Map.put(invalid_attrs, "ends_on", nil)
-      assert {:error, %Ecto.Changeset{} = changeset} = Calendars.create_period(invalid_attrs)
-      assert %{ends_on: ["can't be blank"]} = errors_on(changeset)
-    end
-
-    test "update_period/2 with valid data updates the period" do
-      period = insert(:period)
-      assert {:ok, %Period{} = period} = Calendars.update_period(period, @update_attrs)
-      assert period.html_class == "white"
-      assert period.is_public_holiday == true
-    end
-
-    test "update_period/2 with invalid data returns error changeset" do
-      period = insert(:period)
-      assert {:error, %Ecto.Changeset{}} = Calendars.update_period(period, @invalid_attrs)
-    end
-
-    test "delete_period/1 deletes the period" do
-      period = insert(:period)
-      assert {:ok, %Period{}} = Calendars.delete_period(period)
-      assert_raise Ecto.NoResultsError, fn -> Calendars.get_period!(period.id) end
-    end
-
-    test "change_period/1 returns a period changeset" do
-      period = insert(:period)
-      assert %Ecto.Changeset{} = Calendars.change_period(period)
-    end
-
-    test "find_periods_by_month/2 returns periods if the month has a holiday" do
-      _period_1 = insert(:period, %{starts_on: ~D[2020-02-04], ends_on: ~D[2020-02-07]})
-      _period_2 = insert(:period, %{starts_on: ~D[2020-02-24], ends_on: ~D[2020-02-27]})
-      period_3 = insert(:period, %{starts_on: ~D[2020-04-04], ends_on: ~D[2020-06-07]})
-      periods = Calendars.list_periods()
-      assert Calendars.find_periods_by_month(~D[2020-01-01], periods) == []
-      assert length(Calendars.find_periods_by_month(~D[2020-02-01], periods)) == 2
-      assert [%Period{id: id}] = Calendars.find_periods_by_month(~D[2020-04-01], periods)
-      assert id == period_3.id
-      assert [%Period{id: ^id}] = Calendars.find_periods_by_month(~D[2020-05-01], periods)
-      assert [%Period{id: ^id}] = Calendars.find_periods_by_month(~D[2020-06-01], periods)
-      assert Calendars.find_periods_by_month(~D[2020-08-01], periods) == []
     end
   end
 end
