+++ test/mehr_schulferien/periods_test.exs
@@ -3,8 +3,130 @@ defmodule MehrSchulferien.PeriodsTest do
 
   import MehrSchulferien.Factory
 
-  alias MehrSchulferien.Periods
+  alias MehrSchulferien.{Periods, Periods.Period}
   alias MehrSchulferien.Locations
+
+  describe "periods" do
+    @valid_attrs %{
+      "created_by_email_address" => "george@example.com",
+      "ends_on" => ~D[2010-04-20],
+      "starts_on" => ~D[2010-04-17]
+    }
+    @update_attrs %{
+      "html_class" => "white",
+      "is_public_holiday" => true
+    }
+    @invalid_attrs %{
+      "created_by_email_address" => "george@example.com",
+      "ends_on" => ~D[2010-04-20],
+      "starts_on" => ~D[2010-04-27]
+    }
+
+    test "list_periods/0 returns all periods" do
+      period = insert(:period)
+      assert [period_1] = Periods.list_periods()
+      assert period.id == period_1.id
+    end
+
+    test "get_period!/1 returns the period with given id" do
+      period = insert(:period)
+      assert period_1 = Periods.get_period!(period.id)
+      assert period.id == period_1.id
+    end
+
+    test "create_period/1 with valid data creates a period" do
+      location = insert(:federal_state)
+      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
+
+      valid_attrs =
+        Map.merge(@valid_attrs, %{
+          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
+          "location_id" => location.id
+        })
+
+      assert {:ok, %Period{} = period} = Periods.create_period(valid_attrs)
+
+      assert period.created_by_email_address == "george@example.com"
+      assert period.ends_on == ~D[2010-04-20]
+      assert period.starts_on == ~D[2010-04-17]
+      assert period.html_class == holiday_or_vacation_type.default_html_class
+      assert period.is_public_holiday == holiday_or_vacation_type.default_is_public_holiday
+      assert period.religion_id == holiday_or_vacation_type.default_religion_id
+    end
+
+    test "create_period/1 with attrs overriding holiday_or_vacation_type defaults" do
+      location = insert(:federal_state)
+      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
+
+      valid_attrs =
+        Map.merge(@valid_attrs, %{
+          "location_id" => location.id,
+          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
+          "is_public_holiday" => !holiday_or_vacation_type.default_is_public_holiday
+        })
+
+      assert {:ok, %Period{} = period} = Periods.create_period(valid_attrs)
+      assert period.created_by_email_address == "george@example.com"
+      assert period.is_public_holiday != holiday_or_vacation_type.default_is_public_holiday
+    end
+
+    test "create_period/1 with invalid data returns error changeset" do
+      location = insert(:federal_state)
+      holiday_or_vacation_type = insert(:holiday_or_vacation_type)
+
+      invalid_attrs =
+        Map.merge(@invalid_attrs, %{
+          "holiday_or_vacation_type_id" => holiday_or_vacation_type.id,
+          "location_id" => location.id
+        })
+
+      assert {:error, %Ecto.Changeset{} = changeset} = Periods.create_period(invalid_attrs)
+
+      assert %{starts_on: ["starts_on should be less than or equal to ends_on"]} =
+               errors_on(changeset)
+
+      invalid_attrs = Map.put(invalid_attrs, "ends_on", nil)
+      assert {:error, %Ecto.Changeset{} = changeset} = Periods.create_period(invalid_attrs)
+      assert %{ends_on: ["can't be blank"]} = errors_on(changeset)
+    end
+
+    test "update_period/2 with valid data updates the period" do
+      period = insert(:period)
+      assert {:ok, %Period{} = period} = Periods.update_period(period, @update_attrs)
+      assert period.html_class == "white"
+      assert period.is_public_holiday == true
+    end
+
+    test "update_period/2 with invalid data returns error changeset" do
+      period = insert(:period)
+      assert {:error, %Ecto.Changeset{}} = Periods.update_period(period, @invalid_attrs)
+    end
+
+    test "delete_period/1 deletes the period" do
+      period = insert(:period)
+      assert {:ok, %Period{}} = Periods.delete_period(period)
+      assert_raise Ecto.NoResultsError, fn -> Periods.get_period!(period.id) end
+    end
+
+    test "change_period/1 returns a period changeset" do
+      period = insert(:period)
+      assert %Ecto.Changeset{} = Periods.change_period(period)
+    end
+
+    test "find_periods_by_month/2 returns periods if the month has a holiday" do
+      _period_1 = insert(:period, %{starts_on: ~D[2020-02-04], ends_on: ~D[2020-02-07]})
+      _period_2 = insert(:period, %{starts_on: ~D[2020-02-24], ends_on: ~D[2020-02-27]})
+      period_3 = insert(:period, %{starts_on: ~D[2020-04-04], ends_on: ~D[2020-06-07]})
+      periods = Periods.list_periods()
+      assert Periods.find_periods_by_month(~D[2020-01-01], periods) == []
+      assert length(Periods.find_periods_by_month(~D[2020-02-01], periods)) == 2
+      assert [%Period{id: id}] = Periods.find_periods_by_month(~D[2020-04-01], periods)
+      assert id == period_3.id
+      assert [%Period{id: ^id}] = Periods.find_periods_by_month(~D[2020-05-01], periods)
+      assert [%Period{id: ^id}] = Periods.find_periods_by_month(~D[2020-06-01], periods)
+      assert Periods.find_periods_by_month(~D[2020-08-01], periods) == []
+    end
+  end
 
   describe "periods for certain time frame" do
     setup [:add_federal_state, :add_periods]
